Admin Mode in Book List

+++ WebContent/bookList.jsp
@@ -4,36 +4,48 @@
 <%@ page import="data.dbConnect.DBConnectionPool" %>
 <%@ page import="BookMgmt.Book.*" %>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+
+<%
+	Cookie[] cookies = request.getCookies();
+	String firstName = "";
+	String isAdmin = "";
+	if (cookies != null)
+	{
+		for (int i = 0; i< cookies.length; i++)
+		{
+			if (cookies[i].getName().equals("firstName")) { firstName = cookies[i].getValue(); }
+			if (cookies[i].getName().equals("isAdmin")) { isAdmin = cookies[i].getValue(); }
+		}
+	}
+%>
+
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=US-ASCII">
 <title>Book List - The Best Bookstore</title>
 <link rel="stylesheet" href="css/c06.css"/>
+	 <style type="text/css">
+	 	element.style { border-color: rgb(29, 107, 176); }
+	 	div { background-color: lightblue; padding: 5px; }
+	 </style>
 </head>
 <body>
 
+	<div>
+		<h1>THE BEST BOOKSTORE</h1>
 <%
-	Cookie[] cookies = request.getCookies();
-	String userId = "";
-	if (cookies != null)
-	{
-		for (int i = 0; i< cookies.length; i++)
-		{
-			if (cookies[i].getName().equals("userId"))
-			{
-				userId = cookies[i].getValue();
-				System.out.println(cookies[i].getValue());
-			}
-		}
-	}
-				
-	if (userId.equals(""))
-	{
-		userId = "No Cookies Found";
-	}
-			
+	if (!firstName.equals("")) {%>
+		Welcome <%= firstName %>
+		<% if (isAdmin.equals("true")) {%>[ADMIN MODE]<%}%>
+		<br/><a href="SignOut">Logout</a>
+		
+	<%} else {%>
+		Please <a href="SignIn.jsp">SignIn</a> or <a href="signUpForm_User.jsp">Create an account</a>
+	<%}
 %>
-<p>Username: <%=userId %> </p>
+
+	</div>
+
 	<table>
 		<thead>
 			<tr>	
@@ -73,9 +85,14 @@
 			out.println("<td>"+bookList.get(i).getIsbn13()+"</td>");
 			out.println("<td>"+bookList.get(i).getPrice()+"</td>");
 			out.println("<td>"+bookList.get(i).getInvQty()+"</td>");
-			out.println("<td><a href=\"BookView.jsp?BookId="+bookList.get(i).getBookId()+"\">[View]</a>"+
-				"<br><a href=\"bookUpdateForm.jsp?BookId="+bookList.get(i).getBookId()+"\">[Update]</a>"+
-				"<br><a href=\"DeleteBook?BookId="+bookList.get(i).getBookId()+"\">[Delete]</a></td>");
+			out.println("<td><a href=\"BookView.jsp?BookId="+bookList.get(i).getBookId()+"\">[View]</a>");
+			
+			if (isAdmin.equals("true"))
+			{
+				out.println("<br><a href=\"bookUpdateForm.jsp?BookId="+bookList.get(i).getBookId()+"\">[Update]</a>"+
+						"<br><a href=\"DeleteBook?BookId="+bookList.get(i).getBookId()+"\">[Delete]</a></td>");	
+			}
+
 			out.println("</tr>");
 		}
